Issue 45795: Add simple mechanism to persist usage metrics

+++ api/src/org/labkey/api/usageMetrics/ClientSideMetricService.java
@@ -13,7 +13,7 @@ public interface ClientSideMetricService
         ServiceRegistry.get().registerService(ClientSideMetricService.class, impl);
     }
 
-    int increment(String featureArea, String metricName);
+    long increment(String featureArea, String metricName);
 
     void registerUsageMetrics(String moduleName);
 }
